Break up wrappers.h
