+++ interfaces/cython/cantera/_cantera.pyx
@@ -28,6 +28,7 @@ def bootstrap_cython_submodules():
 
 bootstrap_cython_submodules()
 
+# Import the contents of the individual .pyx files
 from ._onedim import *
 from ._utils import *
 from .base import *
