+++ interfaces/cython/cantera/transport.pxd
@@ -51,6 +51,25 @@ cdef extern from "cantera/transport/TransportData.h" namespace "Cantera":
         double dispersion_coefficient
         double quadrupole_polarizability
 
+cdef extern from "cantera/cython/transport_utils.h":
+    cdef void tran_getMixDiffCoeffs(CxxTransport*, double*) except +translate_exception
+    cdef void tran_getMixDiffCoeffsMass(CxxTransport*, double*) except +translate_exception
+    cdef void tran_getMixDiffCoeffsMole(CxxTransport*, double*) except +translate_exception
+    cdef void tran_getThermalDiffCoeffs(CxxTransport*, double*) except +translate_exception
+    cdef void tran_getSpeciesViscosities(CxxTransport*, double*) except +translate_exception
+    cdef void tran_getMobilities(CxxTransport*, double*) except +translate_exception
+
+    cdef void tran_getMultiDiffCoeffs(CxxTransport*, size_t, double*) except +translate_exception
+    cdef void tran_getBinaryDiffCoeffs(CxxTransport*, size_t, double*) except +translate_exception
+
+    cdef void tran_getViscosityPolynomial(CxxTransport*, size_t, double*) except +translate_exception
+    cdef void tran_getConductivityPolynomial(CxxTransport*, size_t, double*) except +translate_exception
+    cdef void tran_getBinDiffusivityPolynomial(CxxTransport*, size_t, size_t, double*) except +translate_exception
+
+    cdef void tran_setViscosityPolynomial(CxxTransport*, size_t, double*) except +translate_exception
+    cdef void tran_setConductivityPolynomial(CxxTransport*, size_t, double*) except +translate_exception
+    cdef void tran_setBinDiffusivityPolynomial(CxxTransport*, size_t, size_t, double*) except +translate_exception
+
 cdef class GasTransportData:
     cdef shared_ptr[CxxTransportData] _data
     cdef CxxGasTransportData* data
